+++ src/api/groups.rs
@@ -15,6 +15,7 @@ mod group;
 mod groups;
 pub mod members;
 pub mod milestones;
+pub mod projects;
 
 pub use create::BranchProtection;
 pub use create::CreateGroup;